Add menu items only if plugin is enabled. (#12)
* Add menu items only if plugin is enabled.

For some sites giving unrestricted access to LLMs via a central API-Key is not acceptable. So disabling the plugin per default and enabling from time to time is definitely an option.

Unfortunately the plugin does not honor its status (enabled/disabled) - which is fixed by this commit:

Menu items for this plugin are only added if the plugin is enabled.

* Guard index.php when plugin is disabled.

This is a pragmatic and very ugly solution to guard the index page against saved URLs.

Without this guard even when the plugin is disabled, a trainer can use the plugin if he remembers the URL.

* Translate message to English.